+++ scripts/editor/dialog/index.js
@@ -1,8 +1,10 @@
 define(['editor/dialog/img',
-    'editor/dialog/color'
-], function (img, color) {
+    'editor/dialog/color',
+    'editor/dialog/demo'
+], function (img, color, demo) {
     return {
         img: img,
-        color: color
+        color: color,
+        demo: demo
     };
 });
