+++ scripts/editor/type.js
@@ -1,4 +1,4 @@
-define(['data', 'status', 'editor/position'], function (data, status, positionManager) {
+﻿define(['data', 'status', 'editor/position'], function (data, status, positionManager) {
     var layer = $('#type-layer');
 
     var typeMap = {
@@ -6,8 +6,8 @@ define(['data', 'status', 'editor/position'], function (data, status, positionMa
         title: [{name: 'text', title: '字'}],
         subtitle: [{name: 'text', title: '字'}],
         subtitle2: [{name: 'text', title: '字'}],
-        content: [{name: 'text', title: '字'}, {name: 'img', title: '图'}],
-        content2: [{name: 'text', title: '字'}, {name: 'img', title: '图'}]
+        content: [{name: 'text', title: '字'}, {name: 'img', title: '图'}, {name: 'demo', title: '码'}],
+        content2: [{name: 'text', title: '字'}, {name: 'img', title: '图'}, {name: 'demo', title: '码'}]
     };
 
     var currentType;
