+++ scripts/editor/widget/index.js
@@ -1,13 +1,15 @@
 define(['editor/widget/slide',
     'editor/widget/blank',
     'editor/widget/text',
-    'editor/widget/img'
-], function (Slide, Blank, Text, Img) {
+    'editor/widget/img',
+    'editor/widget/demo'
+], function (Slide, Blank, Text, Img, Demo) {
     return {
         'slide': Slide,
         '': Blank,
         'text': Text,
         'title': Text,
-        'img': Img
+        'img': Img,
+        'demo': Demo
     };
 });
